Fold Python build backend into runtime module
